pyproject.toml
[project]
name = "umap-learn-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.1",
    "umap-learn>=0.5.7",
]
